Support auto_tag with dry-run when no tag will be created
This safely allows to check larger parts of the build
earlier (e.g. during a pull request).

An original version did not introduce an additional flag but changed
the existing behavior. However it was rejected because it would break
builds relying on non-PR secrets to pull the base images for the
build.